Make build

+++ massa-models/src/config/constants.rs
@@ -202,7 +202,7 @@ pub const MAX_PARAMETERS_SIZE: u32 = 10_000_000;
 /// Maximum length of `rng_seed` in thread cycle
 pub const MAX_RNG_SEED_LENGTH: u32 = PERIODS_PER_CYCLE.saturating_mul(THREAD_COUNT as u64) as u32;
 /// Maximum number of exports for a smart contract module
-pub const MAX_RUNTIME_MODULE_EXPORTS: u32 = 500; // TODO: Calibrate
+pub const MAX_RUNTIME_MODULE_EXPORTS: usize = 500; // TODO: Calibrate
 
 // ***********************
 // Bootstrap constants
